% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison_data.R
\name{get_all_compdata}
\alias{get_all_compdata}
\title{A list of all compdata suitable for use in gcamdata's old/new tests.}
\usage{
get_all_compdata()
}
\value{
The list of all COMPDATA
}
\description{
This is a list object where [[output_name]] <- compressed(output_data).  In most
cases users should use \code{get_comparison_data} so they do not need to worry about
these details.
}
