% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison_data.R
\name{get_comparison_data}
\alias{get_comparison_data}
\title{Get the comparison data for the given data name}
\usage{
get_comparison_data(name)
}
\arguments{
\item{name}{The name of the comparison data to get}
}
\value{
The comparison data if available, or \code{NULL} otherwise
}
\description{
This method will lookup the data from \code{COMPDATA} and uncompress if it
exists, or return \code{NULL} otherwise.
}
