% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NUM_BINS}
\alias{NUM_BINS}
\title{The number of bins used internally to store \code{COMPDATA}}
\format{An integer}
\usage{
NUM_BINS
}
\description{
We need to store the COMPDATA internally in bins to avoid having to commit any single Rdata that is _too_ big.
The number of bins is calculated as the minimum required so the total size of the data while staying below the max
}
\keyword{datasets}
